Fixes and some refactoring
- Extracted SQLConnector subclasses into separate classes, making it possible to turn SQLConnector into an abstract class.
- Moved db specific syntax workarounds out of BiGGDB and into an abstract method in SQLConnector.
- Fixed pom, disabling building of default-jar. All builds should now properly contain all necessary resources.
- Added a bit of documentation for the initialization of the SQLite db and moved the db into /resources/edu/ucsd/sbrg/bigg.
