+++ .gitignore
@@ -2,9 +2,9 @@
 .settings/
 .classpath
 .project
-/bin/
 /target/
-/resources/bigg.sqlite
+/resources/edu/ucsd/sbrg/bigg/bigg.sqlite
+lib.zip
 /scripts/ModelPolisherTemplate.*
 ModelPolisher.iml
 jsbml.log
