+++ scripts/configureSQLiteDB.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-unzip -od ../resources ../resources/bigg.zip && python createIndices.py
+unzip -od ../resources/edu/ucsd/sbrg/bigg ../resources/edu/ucsd/sbrg/bigg/bigg.zip && python createIndices.py
