+++ src/edu/ucsd/sbrg/bigg/SQLConnector.java
@@ -15,201 +15,26 @@
 package edu.ucsd.sbrg.bigg;
 
 import static edu.ucsd.sbrg.bigg.ModelPolisher.mpMessageBundle;
-import static java.text.MessageFormat.format;
 
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.Properties;
 import java.util.logging.Logger;
-
-import org.sbml.jsbml.util.StringTools;
-import org.sqlite.SQLiteConfig;
-import org.sqlite.SQLiteOpenMode;
 
 /**
  * PostgreSQL database helper class.
  * 
  * @author Andreas Dr&auml;ger
  */
-public class SQLConnector {
+public abstract class SQLConnector {
 
+  protected static Connection connection;
   /**
    * A {@link Logger} for this class.
    */
-  private static Logger logger =
-    Logger.getLogger(PostgreSQLConnector.class.getName());
-  /**
-   *
-   */
-  private static Connection connection;
-
-  /**
-   * @author Andreas Dr&auml;ger
-   */
-  private static enum Keys {
-                            /**
-                             *
-                             */
-                            databaseName,
-                            /**
-                             *
-                             */
-                            dbms,
-                            /**
-                             *
-                             */
-                            host,
-                            /**
-                             *
-                             */
-                            password,
-                            /**
-                             *
-                             */
-                            portNumber,
-                            /**
-                             *
-                             */
-                            server_name,
-                            /**
-                             *
-                             */
-                            user;
-  }
-
-  public class PostgreSQLConnector extends SQLConnector {
-
-    /**
-     *
-     */
-    private Properties properties;
-
-
-    /**
-     * @return
-     * @throws SQLException
-     */
-    public Connection connect() throws SQLException {
-      if (isConnected()) {
-        connection.close();
-      }
-      String url = "jdbc:" + properties.getProperty(Keys.dbms.toString())
-        + "://" + getHost() + ":" + getPort() + "/" + getDatabaseName();
-      connection = DriverManager.getConnection(url, properties);
-      connection.setCatalog(
-        properties.getProperty(Keys.databaseName.toString()));
-      logger.info(format(mpMessageBundle.getString("PSQL_CONNECTED"), getHost(),
-        getPort(), getDatabaseName()));
-      return connection;
-    }
-
-
-    /**
-     * @return
-     */
-    public String getDatabaseName() {
-      return getProperty(Keys.databaseName);
-    }
-
-
-    /**
-     * @return
-     */
-    public String getHost() {
-      return getProperty(Keys.host);
-    }
-
-
-    /**
-     * @return
-     */
-    public int getPort() {
-      String port = getProperty(Keys.portNumber);
-      return (port != null) ? Integer.parseInt(port) : -1;
-    }
-
-
-    /**
-     * @return the properties
-     */
-    protected Properties getProperties() {
-      return properties;
-    }
-
-
-    /**
-     * @param key
-     * @return
-     */
-    private String getProperty(Keys key) {
-      return properties != null ? properties.getProperty(key.toString()) : null;
-    }
-
-
-    /**
-     * @return
-     */
-    public String getUser() {
-      return getProperty(Keys.user);
-    }
-
-
-    /**
-     * @param host
-     * @param port
-     * @param user
-     * @param password
-     * @param dbName
-     * @throws ClassNotFoundException
-     */
-    public PostgreSQLConnector(String host, int port, String user,
-      String password, String dbName) throws ClassNotFoundException {
-      Class.forName("org.postgresql.Driver");
-      properties = new Properties();
-      properties.setProperty(Keys.dbms.toString(), "postgresql");
-      properties.setProperty(Keys.host.toString(), host);
-      properties.setProperty(Keys.user.toString(), user);
-      properties.setProperty(Keys.databaseName.toString(), dbName);
-      properties.setProperty(Keys.password.toString(),
-        password != null ? password : "");
-      properties.setProperty(Keys.host.toString(), host);
-      properties.setProperty(Keys.portNumber.toString(),
-        Integer.toString(port));
-      logger.fine(format("{0}@{1}:{2}, password={3}", user, host, port,
-        StringTools.fill(password.length(), '*')));
-    }
-  }
-
-  public class SQLiteConnector extends SQLConnector {
-
-    /**
-     * @return
-     * @throws SQLException
-     */
-    public Connection connect() throws SQLException {
-      if (isConnected()) {
-        connection.close();
-      }
-      SQLiteConfig config = new SQLiteConfig();
-      config.setOpenMode(SQLiteOpenMode.OPEN_MEMORY);
-      connection =
-        DriverManager.getConnection("jdbc:sqlite::resource:bigg.sqlite");
-      config.apply(connection);
-      logger.info(mpMessageBundle.getString("SQLITE_CONNECTED"));
-      return connection;
-    }
-
-
-    /**
-     * @throws ClassNotFoundException
-     */
-    public SQLiteConnector() throws ClassNotFoundException {
-      Class.forName("org.sqlite.JDBC");
-    }
-  }
+  private static final transient Logger logger =
+    Logger.getLogger(SQLConnector.class.getName());
 
 
   /**
@@ -221,6 +46,9 @@ public class SQLConnector {
       logger.fine(mpMessageBundle.getString("CONNECTION_CLOSED"));
     }
   }
+
+
+  public abstract Connection connect() throws SQLException;
 
 
   /*
@@ -264,4 +92,10 @@ public class SQLConnector {
     Statement stmt = connection.createStatement();
     return stmt.executeQuery(String.format(query, args));
   }
+
+
+  /**
+   * @return
+   */
+  public abstract String selectConcat();
 }
